Refactor: Nebulae family: make casts when converting between Rack frames and Clouds shortframes explicit.

+++ src/etesia.cpp
@@ -352,8 +352,10 @@ struct Etesia : SanguineModule {
 				   between the input and output SRC.
 				*/
 				for (int block = 0; block < outputLength; ++block) {
-					input[block].l = clamp(convertedFrames[block].samples[currentChannel] * 32767.f, -32768.f, 32767.f);
-					input[block].r = clamp(convertedFrames[block].samples[currentChannel + 1] * 32767.f, -32768.f, 32767.f);
+					input[block].l = static_cast<short>(
+						clamp(convertedFrames[block].samples[currentChannel] * 32767.f, -32768.f, 32767.f));
+					input[block].r = static_cast<short>(
+						clamp(convertedFrames[block].samples[currentChannel + 1] * 32767.f, -32768.f, 32767.f));
 				}
 
 				etesiaParameters[channel] = etesiaProcessors[channel]->mutable_parameters();
@@ -437,8 +439,8 @@ struct Etesia : SanguineModule {
 
 				// Convert output buffer.
 				for (size_t block = 0; block < etesia::kMaxBlockSize; ++block) {
-					renderedFrames[block].samples[currentChannel] = output[block].l / 32768.f;
-					renderedFrames[block].samples[currentChannel + 1] = output[block].r / 32768.f;
+					renderedFrames[block].samples[currentChannel] = static_cast<float>(output[block].l / 32768.f);
+					renderedFrames[block].samples[currentChannel + 1] = static_cast<float>(output[block].r / 32768.f);
 				}
 			}
